requirements.txt
asgiref==3.2.7
boto3==1.12.31
botocore==1.15.31
Django==3.0.4
django-crequest==2018.5.11
django-crispy-forms==1.9.0
django-recaptcha==2.0.6
django-sitemaps==1.1.1
django-summernote==0.8.11.6
docutils==0.15.2
jmespath==0.9.5
lxml==4.5.0
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.8
